requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0

# Logging and monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Video processing
decord>=0.6.0
av>=10.0.0

# Model training dependencies
transformers==4.49.0
accelerate>=0.24.0
peft>=0.6.0  # For LoRA training
bitsandbytes>=0.41.0  # For quantization
sentencepiece>=0.1.99  # For tokenizer
einops>=0.7.0  # For tensor operations

# Flash Attention (requires CUDA and proper compilation)
# flash-attn>=2.3.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
ipython>=8.14.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# git+https://github.com/huggingface/transformers
# pip install -U git+https://github.com/huggingface/transformers
